requirements.txt
asgiref==3.7.2
astunparse==1.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Django==4.2.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-yasg==1.21.6
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flatbuffers==23.3.3
gast==0.4.0
gevent==22.10.2
geventhttpclient==2.0.9
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
h5py==3.8.0
idna==3.4
imageio==2.5.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
keras==2.10.0
keras-nlp==0.4.1
Keras-Preprocessing==1.1.2
libclang==16.0.0
locust==2.15.1
MarkupSafe==2.1.2
msgpack==1.0.5
networkx==3.1
nltk==3.4.5
opt-einsum==3.3.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
psutil==5.9.5
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.0
python-dateutil==2.8.2
python-decouple
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
roundrobin==0.0.4
scikit-image==0.15.0
six==1.16.0
sqlparse==0.4.4
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.10.0
termcolor==2.2.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.1.7
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
